Remove useless messages; add archive ignoring

+++ butler.py
@@ -30,37 +30,40 @@ files_extension = [
     ".allow",
 ]
 
+archives_extension = [".zip", ".7z", ".gz", ".bz", ".gzip", ".bzip"]
+
 target_dir_name = "ALL"
 
 
-def clean_the_dir(directory_path: str):
-    for files in os.listdir(directory_path):
-        path = os.path.join(directory_path, files)
-        if "butler" in (path.split("/")[-1]):
-            print("Skipped " + path.split("/")[-1])
-        else:
-            try:
-                shutil.rmtree(path)
-            except OSError:
-                os.remove(path)
+def clean_the_dir(path_to_clean: str):
+    if args.clean == "/":
+        print("It is totally not great idea to remove all things")
+        exit(1)
+    else:
+        for files in os.listdir(path_to_clean):
+            path = os.path.join(path_to_clean, files)
+            if "butler" in (path.split("/")[-1]):
+                print("Skipped " + path.split("/")[-1])
+            else:
+                try:
+                    shutil.rmtree(path)
+                except OSError:
+                    os.remove(path)
 
 
 def group_up_files(new_dir_name: str):
-    print(new_dir_name)
     for file in os.listdir(args.dir):
         for ext in files_extension:
             if file.endswith(ext):
                 file_path = os.path.join(args.dir, file)
-                print(file_path)
                 if args.dir == ".":
                     new_dir_path = new_dir_name + ext.upper()
                 else:
                     new_dir_path = args.dir + new_dir_name + ext.upper()
-                print(new_dir_path)
                 try:
                     os.mkdir(new_dir_path)
-                except OSError as os_error:
-                    print(os_error)
+                except OSError:
+                    pass
                 shutil.move(file_path, new_dir_path)
 
 
@@ -70,11 +73,12 @@ def create_archive(dir_to_archive: str):
     with ZipFile(str(date_time) + ".zip", "w") as zip_obj:
         for folder_name, sub_folders, filenames in os.walk(dir_to_archive):
             for filename in filenames:
-                if ".zip" in filename:
-                    pass
-                else:
-                    zip_path = os.path.join(folder_name, filename)
-                    zip_obj.write(zip_path, basename(zip_path))
+                for a_ext in archives_extension:
+                    if filename.endswith(a_ext):
+                        pass
+                    else:
+                        zip_path = os.path.join(folder_name, filename)
+                        zip_obj.write(zip_path, basename(zip_path))
 
 
 parser = argparse.ArgumentParser(
